Add Role Based Authorization
- AuthHandler now has a parameter called "role". The user requesting
a resource must have permissions greater than or equal to the role
which is specified in the AuthHandler in the route definition.

A couple of examples:

For POST /api/users let the AuthHandler declaration be:
AuthHandler{env, UsersCreate, "system_admin"}
A user with role researcher tries to POST /api/users. 401 unauthorized
is returned because they are not allowed to access the resource.

For GET /api/readings, let the AuthHandler declaration be:
AuthHandler{env, ReadingsIndex, "researcher"}
A user with role system_admin tries to GET /api/readings. They are
allowed to because they have a role which is greater than or equal
to the "researcher" role.