Disable archiving of sensor types
It probably doesn't make sense to archive a sensor type. There will be some
undefined behaviour if a sensor type attached to a buoy is archived.
Moreover, how do we handle display of (archived types of) sensors attached
to a buoy? Do we allow adding of warning triggers for archived sensor
types? Do we display archived sensor types on the dashboard?